% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{config_game}
\alias{config_game}
\title{Configure wordPuzzleR}
\usage{
config_game(..., .verbose = TRUE)
}
\arguments{
\item{...}{Arguments passed on to configurations. Valid names may be:
\describe{
\item{dict}{(String) path to dictionary file, where each line is a word.}
\item{min_len}{(Integer) minimal word length, default 3.}
\item{max_len}{(Integer) maximal word length, default 8.}
\item{guess}{(Integer) maximal guesses, default 10.}
\item{pattern}{(String) Regular expression to filter qualified words,
default "[A-Za-z]+".}
}}

\item{.verbose}{(Logical) whether config messages should be printed.}
}
\value{
Named list of new configurations, invisibly.
}
\description{
\code{config_game} configures wordPuzzleR, or show current configuration
when used with no arguments.
}
\examples{
# Show current config
config_game()
}
