% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordPuzzleR-package.R
\docType{package}
\name{wordPuzzleR-package}
\alias{wordPuzzleR}
\alias{wordPuzzleR-package}
\title{wordPuzzleR: Word Puzzle Game in R}
\description{
The word puzzle game requires you to find out the letters in a word within a limited number of guesses. In each round, if your guess hit any letters in the word, they reveal themselves. If all letters are revealed before your guesses run out, you win this game; otherwise you fail. You may run multiple games to guess different words.
}
\examples{
# Run word puzzle game
run_game()
}
\author{
\strong{Maintainer}: Xiurui Zhu \email{zxr6@163.com}

}
\keyword{internal}
